refactor project file structure
