+++ src/res/img/jira-icon.png
