+++ src/ticket/PriorityTile.js
@@ -1,33 +1,36 @@
 import React, { useState } from 'react';
 import styled from 'styled-components';
-import useComponentVisible from '../customHooks/useComponentVisible';
+import useComponentVisible from '../common/customHooks/useComponentVisible';
 
-import BookmarkIcon from '@mui/icons-material/Bookmark';
-import CheckBoxIcon from '@mui/icons-material/CheckBox';
-import BugReportIcon from '@mui/icons-material/BugReport';
+import ArrowUpwardIcon from '@mui/icons-material/ArrowUpward';
+import ArrowDownwardIcon from '@mui/icons-material/ArrowDownward';
 
 const MAPPING = [
   {
-    title: 'Story',
-    code: 'story',
+    title: 'Highest',
+    code: 'sev2',
   },
   {
-    title: 'Task',
-    code: 'task',
+    title: 'Higher',
+    code: 'sev1',
   },
   {
-    title: 'Bug',
-    code: 'bug',
+    title: 'Medium',
+    code: 'high',
+  },
+  {
+    title: 'Lowest',
+    code: 'low',
   },
 ];
 
-const TaskTypeSelect = (props) => {
+const PriorityTile = (props) => {
   const [isExpanded, setIsExpanded] = useState(false);
-  const value = MAPPING.filter((type) => type.code === props.issueType)[0]
-    .title;
+
+  const value = MAPPING.filter((prio) => prio.code === props.priority)[0].title;
 
   const onClick = (newVal) => {
-    props.setIssueType(newVal);
+    props.setPriority(newVal);
     setIsExpanded(false);
   };
 
@@ -38,9 +41,7 @@ const TaskTypeSelect = (props) => {
         style={{ border: '1px solid #dfe1e6' }}
       >
         <UserInfo>
-          {RenderTicketTypeIcon(
-            MAPPING.filter((type) => type.title === value)[0].code
-          )}
+          {RenderTicketSeverityIcon(value)}
           {value}
         </UserInfo>
       </Container>
@@ -55,17 +56,17 @@ const DropDown = (props) => {
   return (
     <DropDownComponent ref={ref}>
       {isComponentVisible &&
-        MAPPING.map((type, i) => {
-          if (type.title !== props.value) {
+        MAPPING.map((prio, i) => {
+          if (prio.title !== props.value) {
             return (
               <Container
                 key={i}
-                onClick={() => props.onClick(type.code)}
+                onClick={() => props.onClick(prio.code)}
                 style={{ border: '1px solid #F4F5F7' }}
               >
                 <UserInfo>
-                  {RenderTicketTypeIcon(type.code)}
-                  {type.title}
+                  {RenderTicketSeverityIcon(prio.title)}
+                  {prio.title}
                 </UserInfo>
               </Container>
             );
@@ -77,22 +78,38 @@ const DropDown = (props) => {
   );
 };
 
-const RenderTicketTypeIcon = (type) => {
+const RenderTicketSeverityIcon = (priority) => {
+  var color = '';
   const fontSize = 18;
-
-  switch (type) {
-    case 'story':
-      return <BookmarkIcon sx={{ color: '#65ba43', fontSize: { fontSize } }} />;
-    case 'task':
-      return <CheckBoxIcon color="primary" sx={{ fontSize: { fontSize } }} />;
-    case 'bug':
-      return <BugReportIcon color="action" sx={{ fontSize: { fontSize } }} />;
+  switch (priority) {
+    case 'Highest':
+      color = '#cd1316';
+      break;
+    case 'Higher':
+      color = '#e97f33';
+      break;
+    case 'Medium':
+      color = '#57a55a';
+      break;
+    case 'Lowest':
+      color = '#2d8738';
+      break;
     default:
-      return <BookmarkIcon sx={{ color: '#65ba43', fontSize: { fontSize } }} />;
+      color = '#2d8738';
+  }
+
+  if (priority === 'Highest' || priority === 'Higher') {
+    return (
+      <ArrowUpwardIcon sx={{ color: { color }, fontSize: { fontSize } }} />
+    );
+  } else {
+    return (
+      <ArrowDownwardIcon sx={{ color: { color }, fontSize: { fontSize } }} />
+    );
   }
 };
 
-export default TaskTypeSelect;
+export default PriorityTile;
 
 const DropDownComponent = styled.div`
   background-color: rgb(244 245 247);
