+++ src/ticket/StatusTile.js
@@ -1,6 +1,6 @@
 import React, { useState } from 'react';
 import styled from 'styled-components';
-import useComponentVisible from '../customHooks/useComponentVisible';
+import useComponentVisible from '../common/customHooks/useComponentVisible';
 
 const MAPPING = [
   {
